Release notes
```noteworthy operator
On AWS the hosted zones potentially usable by a set of credentials
are now filtered by the permissions for reading their record sets.
So, even if the list permission is given for all zones, only zones
whose record sets can really be read will be added to the available
zone list.
```

This avoids permission errors, if the actual credentials do not allow
reading of record sets for all the zones returned by the list operation.
This cannot be avoided by the standard filter options for zones and
domains offerend in the provider objects, because this filtering
is done on a provider-basis after the evaluation at the account level
common for all providers using the same credential set.