Add a Netlify DNS provider
Add a DNS provider that uses the Netlify API to provision DNS
records. Authentication is based on personal access tokens:
https://docs.netlify.com/accounts-and-billing/user-settings/#connect-with-other-applications

Netlify does not provide a record update API, so the update
operation is implemented as a combination of delete and create.

This PR is based on the work previously done to add a Cloudflare
DNS provider: https://github.com/gardener/external-dns-management/pull/60

Update the helm chart accordingly.

Fixes #146

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>